# First iteration of perf improvements