This is the first iteration of improving perf for Java API view parser.

- Updated the JavaASTAnalyser to use parallelstream() for scanning types. 
- Using local schema file for validation instead of downloading the schema each time 

Used **ARM Compute library** for testing this as it has quite a large API surface.
**Before:**
Total time for 5 iterations: 27.283 seconds
Average time per iteration: 5.457 seconds

**After:**
Total time for 5 iterations: 19.043 seconds
Average time per iteration: 3.809 seconds